feat: add blockchain API provider
