+++ relay_rpc/src/auth/cacao/signature/eip1271/get_rpc_url.rs
@@ -1,0 +1,5 @@
+use url::Url;
+
+pub trait GetRpcUrl {
+    fn get_rpc_url(&self, chain_id: String) -> Option<Url>;
+}
